requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Authlib==1.4.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.0.1
